+++ packages/theme_tailor_annotation/lib/src/tailor_theme_extension.dart
@@ -2,9 +2,9 @@ import 'package:meta/meta_meta.dart';
 
 /// Hints generator that the type conforms to ThemeExtension and can call lerp()
 /// on it.
-const themeExtension = _ThemeExtension();
+const themeExtension = TailorThemeExtension();
 
 @Target({TargetKind.field})
-class _ThemeExtension {
-  const _ThemeExtension();
+class TailorThemeExtension {
+  const TailorThemeExtension();
 }