+++ packages/theme_tailor/CHANGELOG.md
@@ -1,6 +1,8 @@
 # 3.1.1
+- Support analyzer: ">=7.5.9 <9.0.0"
 - Support build 3.0.0 and ^4.0.0
 - Support source_gen 3.0.0 and ^4.0.0
+- Use the same versioning for other theme_tailor packages
 
 # 3.1.0
 - Support build 3.0.0.