+++ packages/theme_tailor_toolbox/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 3.1.1
 - Compatibility with theme_tailor 3.1.1
+- Use the same versioning for other theme_tailor packages
 
 # 3.1.0
 - Require Dart 3.8.0